requirements.txt
fastapi==0.115.12
python-dotenv==1.1.0
Telethon==1.39.0
uvicorn==0.34.1
